:truck: refactored code organization
